update fix bug delete path image

+++ operations/delete_entity.py
@@ -1,9 +1,10 @@
+import os
+from logs import LOGGER
+from config import DEFAULT_TABLE
 import sys
 
 sys.path.append("..")
-from config import DEFAULT_TABLE
-from logs import LOGGER
-import os
+
 
 def do_drop_entity(table_name, milvus_cli, mysql_cli, name_folder):
     if not table_name:
@@ -11,12 +12,14 @@ def do_drop_entity(table_name, milvus_cli, mysql_cli, name_folder):
     try:
         if not milvus_cli.has_collection(table_name):
             return f"Milvus doesn't have a collection named {table_name}"
-        milvus_ids, path_images = mysql_cli.search_milvus_ids_by_name_folder(table_name, name_folder)
+        milvus_ids, path_images = mysql_cli.search_milvus_ids_by_name_folder(
+            table_name, name_folder)
         if milvus_ids:
             status = milvus_cli.delete_entity(table_name, milvus_ids)
             mysql_cli.delete_entity(table_name, milvus_ids)
             for path_image in path_images:
-                os.remove(path_image)
+                if os.path.exists:
+                    os.remove(path_image)
             os.rmdir("/".join(path_image.split("/")[:-1]))
             return status
     except Exception as e:
